API references

SearchSpaces.cardinality — Function

cardinality(searchspace)

Cardinality of the search space.

Example

julia> cardinality(Permutations(5))
120

julia> cardinality(Bounds(lb = zeros(2), ub = ones(2)))
Inf

julia> cardinality(Bounds(lb = zeros(Int, 2), ub = ones(Int,2)))
4

julia> mixed = MixedSpace(
                          :W => Categorical([:red, :green, :blue]),
                          :X => Permutations(3),
                          :Y => BitArrays(3),
                         );

julia> cardinality(mixed)
144

SearchSpaces.isinbounds — Function

isinbounds(item, searchspace) --> Bool

Determine whether an item is in the given searchspace.

SearchSpaces.ispermutation — Function

ispermutation(item, searchspace) --> Bool

Determine whether an item is in the given searchspace.

Missing docstring.

Missing docstring for ispermutation. Check Documenter's build log for details.

SearchSpaces.isinspace — Function

isinspace(item, searchspace) --> Bool

Determine whether an item is in the given searchspace.

See also in.

SearchSpaces.Variable — Type

Variable(name, searchspace)

A structure to define a variable in the search space: searchspace.

SearchSpaces.@var — Macro

@var

A macro to define a Variable.