API references

Search Spaces

The search spaces can be defined as follows.

$k$-Permutation Space

SearchSpaces.Permutations — Type

Permutations(values; k)
Permutations(k)

Define a search space defined by permuting the values of size k (k-permutations).

Categorical

SearchSpaces.Categorical — Function

Categorical(categories)

Define a search space given by the provided categories (Vector).

Bit Arrays

SearchSpaces.BitArrays — Type

BitArrays(;dim)

Define a search space delimited by bit arrays.

Bounds (Box-space)

Bounds

Built-in Samplers

SearchSpaces.Grid — Type

Grid(searchspace; npartitions)

Return an iterator over the given searchspace.

The npartitions controls the number of partitions for each axis when searchspace isa Bounds (3 by default).

Examples

julia> for x in Grid(Permutations([:red, :green, :blue]))
           @show x
       end
x = [:red, :green, :blue]
x = [:red, :blue, :green]
x = [:green, :red, :blue]
x = [:green, :blue, :red]
x = [:blue, :red, :green]
x = [:blue, :green, :red]

julia> for x in Grid(Bounds(lb=[-1.0, -1], ub=[1, 0.0]), npartitions=3)
           @show x
       end
x = [-1.0, -1.0]
x = [0.0, -1.0]
x = [1.0, -1.0]
x = [-1.0, -0.5]
x = [0.0, -0.5]
x = [1.0, -0.5]
x = [-1.0, 0.0]
x = [0.0, 0.0]
x = [1.0, 0.0]

julia> mixed = MixedSpace(
                                 :W => Categorical([:red, :green, :blue]),
                                 :X => Permutations(2),
                                 :Y => BitArrays(2),
                                );

julia> collect(Grid(mixed))
24-element Vector{Any}:
 Dict{Symbol, Any}(:W => :red, :X => [1, 2], :Y => Bool[0, 0])
 Dict{Symbol, Any}(:W => :green, :X => [1, 2], :Y => Bool[0, 0])
 Dict{Symbol, Any}(:W => :blue, :X => [1, 2], :Y => Bool[0, 0])
 Dict{Symbol, Any}(:W => :red, :X => [2, 1], :Y => Bool[0, 0])
 Dict{Symbol, Any}(:W => :green, :X => [2, 1], :Y => Bool[0, 0])
 Dict{Symbol, Any}(:W => :blue, :X => [2, 1], :Y => Bool[0, 0])
 ⋮
 Dict{Symbol, Any}(:W => :blue, :X => [2, 1], :Y => Bool[0, 1])
 Dict{Symbol, Any}(:W => :red, :X => [1, 2], :Y => Bool[1, 1])
 Dict{Symbol, Any}(:W => :green, :X => [1, 2], :Y => Bool[1, 1])
 Dict{Symbol, Any}(:W => :blue, :X => [1, 2], :Y => Bool[1, 1])
 Dict{Symbol, Any}(:W => :red, :X => [2, 1], :Y => Bool[1, 1])
 Dict{Symbol, Any}(:W => :green, :X => [2, 1], :Y => Bool[1, 1])
 Dict{Symbol, Any}(:W => :blue, :X => [2, 1], :Y => Bool[1, 1])

SearchSpaces.AtRandom — Type

AtRandom(searchspace;rng)

Define a random iterator for the search space.

Miscellaneous

SearchSpaces.cardinality — Function

cardinality(searchspace)

Cardinality of the search space.

Example

julia> cardinality(Permutations(5))
120

julia> cardinality(Bounds(lb = zeros(2), ub = ones(2)))
Inf

julia> cardinality(Bounds(lb = zeros(Int, 2), ub = ones(Int,2)))
4

julia> mixed = MixedSpace(
                          :W => Categorical([:red, :green, :blue]),
                          :X => Permutations(3),
                          :Y => BitArrays(3),
                         );

julia> cardinality(mixed)
144

SearchSpaces.isinbounds — Function

isinbounds(item, searchspace) --> Bool

Determine whether an item is in the given searchspace.

SearchSpaces.ispermutation — Function

ispermutation(item, searchspace) --> Bool

Determine whether an item is in the given searchspace.

SearchSpaces.isinspace — Function

isinspace(item, searchspace) --> Bool

Determine whether an item is in the given searchspace.

See also in.

SearchSpaces.Variable — Type

Variable(name, searchspace)

A structure to define a variable in the search space: searchspace.

SearchSpaces.@var — Macro

@var x in searchspace

A macro to define a Variable for in the given search space.